LKVS: LIBZBC Updates, CZ Calculation
Added write pointer updating because libzbc no longer supports
this functionality.

We used to have commands to set zones on a prototype drive, so I
have changed liblkvs to skip over a single or multiple conventional
zones at the beginning of the drive. This value used to be hard
coded.

In addition I had to change some testing code and scripts to support
our current prototype drives.

Signed-off-by: Adam Manzanares <adam.manzanares@hgst.com>