SCSI backend: Fix SAT detection
SAT-4 specifies that ZBC_IN (report zones) command translation should be
done so that the ZAC report zone command is sent with a 512B aligned
buffer length (return page count field). However, Linux in-kernel SAT
does not implement this and returns an error. So detect SAT by sending
a report zone with a 512B buffer size instead of 64B.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>