Introduce open flags
Introduce new open flags beyond the access mode flags to allow
controlling the backend driver to use. This is especially interesting
for bypassing any SAT layer with a ZAC drive and force the use of the
ATA backend driver instead of the SCSI backend driver.

For device tests, this also allows disabling the use of the block device
backend driver.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>