ATA Backend: Use SCSI commands for I/O
If we reach the ATA backend driver on open, it means that the system has
no functional ZBC/ZAC SAT layer. However, the system may still be able
to translate read write and flush commands for the ZAC drive. Test this
and enable the use of SCSI commands if correct translation is detected.
On a regular kernel, this will lead to the use of NCQ commands, and so
to higher performance for multi-threaded applications.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>