SG: Support direct I/O
If O_DIRECT is specified with zbc_open(), use SG direct I/Os by setting
the appropriate flag in the SG I/O header structure. Note that the SG
node file itself is not open with O_DIRECT in this case, but commands
sent to the device with the SG_IO ioctl will use direct I/O mapping of
the command buffer.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>