Sync up test scripts with the latest emulator
The emulator has been fixed not to convert Stasis
zones to Empty by Open Zone operation and some of
the test started to fail because of that. These
changes allow the tests to pass on a fully or partially
converted Realms devices. There still are test failures
when running the test suite against a full CMR drive -
this case will be taken care of in the next commit.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>