Use flock instead of pthread_mutex
Using mmap-ed pthread_mutex_t for locking metadata of the fake
driver has some serious drawbacks if the application owning the
lock crashes or the system reboots unexpectedly. In such cases,
the mutex can be left in a locked state and prevent correct
operation of the fake driver.

This patch replaces the mutex with file locking (flock) to avoid
persistency of the lock state across application runs or system
reboot.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>