Use flock instead of pthread_mutex
Using mmap-ed pthread_mutex_t for locking has some serious
drawbacks in this application as the lock may get stuck in
locked state after unexpected reboot or an application crash.

This patch replaces the mutex with the file lock (flock) and
this change makes the code both simpler and more resilient.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>