Add support for MODE SENSE and MODE SELECT
The current DH-SMR spec mandates using these SCSI commands to
get and set some values on the device, the most important being
the number of zones to convert. This commit adds a new libgary
call to issue both of these commands. The next step is to intro
a new utility to read and change the settings via this mechanism.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>