Update cross-zone-type test so it can run
Change 01.062 so that it does not require an active SWR zone contiguous
with the last conventional zone to run the test.  It is enough to cross
the boundary between conventional LBA space and whatever comes after it,
whether active SWR or not.

Also, add 01.064 which runs a similar test, but extends the read to reach
the first active SWR zone.  The read will span any intervening LBAs.