Emulate DOMAIN REPORT if it's not supported by DUT
Also, fixed a few tests to keep in sync with some emulator changes.

With non-convertible domains and zones, some of the zones in the
opposite LBA space will never be activated. The emulator now sets
their condition to Offline rather than Inactive to mark the
difference.

Emulator now sets the start LBA of a non-convertible domain if there
is an attempt to convert it.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>