Add mutations for no DOMAIN REPORT and RO/Offline
Two mutations are added to the list in dev_control utility.

One is to mutate to the device with support of DOMAIN REPORT
absent. libzbc tries to emulate it using REPORT ZONES and
ZONE QUERY.

The other mutation is a device with several readonly and offline
zones. This one is intended to help extent test coverage for
handling these zone conditions.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>