Add test that SWP zones unaffected by OZR check
Also can now check conventional and WPC zones.  The test opens the
maximum number of SWR or SWP zones, then tries to write another zone
of any type.  If the opened zones and the written zone are all SWR,
the write is expected to return "insufficient resources"; else the
write should succeed.