move zero-length write tests to Section 8
Each of these two tests issues a zero-length Write at the write pointer
to an EMPTY or CLOSED zone and checks the resulting zone condition,
which the ZBC standard appears to specify as IMPLICITLY OPENED.  Apart
from changing the zone condition, it would appear that a zero-length
WRITE should (by SBC) successfully write nothing.

However there appears to be an unfortunate interaction between the ZBC
spec and a translation done in some HBAs between SCSI protocol and ATA
protocol: the ATA protocol has no way to specify zero-length I/O -- it
uses zero in its length field to denote the maximum possible value.

So translating HBAs may have firmware that "fakes up" a target response
to return to the initiator, without ever passing the command to the
actual target device.  Probably this trick works fine with regular
SBC/ATA drives because the initiator can't tell this is happening.

But with a ZBC drive a zero-length write has the host-visible effect of
changing the state of the zone -- and the HBA can't really fake that,
but neither can it translate the zero-length write.

The upshot is that when using SCSI to ATA translation the drive never
sees a zero-length write, and the zone condition remains unchanged.

However the tests work properly when used on a real SCSI drive.  So
this checkin moves the two tests to test Section 8 which is reserved
for tests to be run on SCSI drives only

Signed-off-by: David Butterfield <David.Butterfield@wdc.com>