Make zone activation work on domain ID basis
Command format changes to reflect the new spec. The new code
is more complex than before, but it seems more flexible.

One problem that needs to be resolved is zone size information
in REPORT REALMS. Realm activation requires the number of zones
to activate and that value is not possible to deduce just from
start and end LBAs of a realm. I solved this by issuing
REPORT ZONES for the start LBA of every realm in every domain
to obtain the realm zone size, but this approach is not very
scalable - ideally this value should be reported as a part
of REPORT REALMS output.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>