test zone OPs: ZONE-ID ignored when ALL bit is set
Add test scripts 01.127, 01.137, 01.047, 01.057 each of which issues a
zone operation:
	(a) with the ALL bit set, and
	(b) specifying an LBA (ZONE-ID) that is out of range.

The operations are expected to succeed, despite being given an invalid
ZONE-ID, because the zone OPs are specified by T10 ZBC to ignore the
ZONE-ID when the ALL bit is set.

Also fix the test programs and library to ensure that when in test mode,
the op gets through to the device as specified by the script.

Signed-off-by: David Butterfield <David.Butterfield@wdc.com>