tests: Fix expected ASC/ASCQ value
The ZBC and ZAC standards state that specifying an LBA that is not a
zone ID (the first LBA of a zone) for the open, close, finish and reset
commands should result in an illegal request with the ascq set to
"invalid field in cdb". However, in the case of an LBA that is beyond
the device maximum LBA, some implementations return the more natural and
descriptive "LBA out of range" error.

While not being exactly as stated by the standard, accept this behavior
and allow tests of these commands to pass if the LBA out of range error
is returned by the device for a command executed with the specified LBA
beyond the disk maximum LBA.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>