Improve zbc_open() behavior
Prevent opening of a regular block device or regular file if the device
or file was not prepared for use with the fake backend driver using
zbc_set_zones() (or the command line tool zbc_set_zones).

To do that, introduce the internal open flag ZBC_O_SETZONES which if
specified allows opening a regular device or file with the fake backend
driver to execute zbc_set_zones(). Without this flag, attempt at opening
the unprepared device or file always result in an error.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>