tools: Add data pattern check
Introduced the -p option in zbc_write_zone to specify a byte
pattern to be used for writing. If this option is specified
on the command line, the program will fill every byte that is
being written with this value.

The complementary option in zbc_read_zone tells the program
to verify if every read data byte contains the pattern value.
If there is a mismatch, the program will output the location
of the first mismatched byte.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>