lib: always follow symlinks in zbc_open()
If a device mapper is placed on top of a zoned block device,
the resulting named device file (i.e. /dev/mapper/dmxx-yyy)
cannot be used by libzbc as it is a symlink to i.e /dev/dm-0.

Fix this by resolving the real path of any filename passed
to zbc_open() to remove symlinks and "." or ".." componenets.
This allows simplifying zbc_sg.c too.

Signed-off-by: Hans Holmberg <hans.holmberg@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>