Change return code for misaligned i/o
The library checks for sector alignment in preadv/pwritev and fails
the I/O without sending it to the driver. The I/O is failed with
EINVAL error code and this is not what block and SCSI subsystems do -
the I/O is failed by them with EIO error. This may confuse some tests.

This commit changes the returned error code for non-block aligned
I/O from EINVAL to EIO.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>