lib: fix zbc_device_is_zoned()
In zbc_device_is_zoned(), the device path specified iby the user is
resolved using zbc_realpath() but the driver open method is being called
with the unresolved device path, which causes errors if the user
specifies a path with symbolic links.

Fix this by using the resoled path returned by zbc_realpath().

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>