lib: Fix maximum I/O size with unaligned I/O buffers
If libzbc read or write functions are called with a buffer address not
aligned on the system page size, large IO requests may fail with EINVAL
due to the inability of the kernel to map the IO buffer pages within the
boundaries of the device limits (number of SGL segments). libzbc chunks
large IO request into smaller requests within the device capabilities,
but this operation was not accounting for the IO buffer page alignment.

Fix this by counting also the number of pages an IO buffer uses and
reduce further the maximum IO size allowed to ensure that the kernel can
always map the request buffer.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>