tools/zone_op: Correctly display zone number
zbc_zone_op() shows the correct zone number only if zone number
addressing is specified on the command line. When LBA or sector
addressing is used, the zone number is always shown as zero in the
command output.

Fix this by getting a full zone listing and searching for the target
zone using the zone start address as specified (sector or LBA).

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>
Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>