configure: Fix use of AC_CHECK_HEADERS
Many users have reported problems with the detection of the presence and
usability of the files scsi/scsi.h, scsi/sg.h and libgen.h. Since these
files are part of glibc-headers, they can safely be assumed to be
present and the AC_CHECK_HEADERS() macro calls to check for these files
removed, avoiding user problems.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>