tests: Pass if FINISH at max_open succeeds
Two existing tests open MAX OPEN zones and then try to finish another
zone. These tests assume that the drive would have to open the zone
internally before finishing and that would cause an OZR error.
Depending on the actual drive design, this may not always be the case
and a drive may finish the zone successfully.

This patch allows the successful finish operation outcome to pass
the test.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>