rpm: Add source rpm build
Modify the "make rpm" build directive to also generate a source rpm.

To enable this rpm to be rebuilt outside of libzbc source tree,
dynamically generate the libzbc spec file to hard code the version
number to avoid unresolved macro problems. Also remove the clean section
and the archive macro. The source file name is generated using the name
and version macros already defined in the spec file.

While at it, also add the rpm groups.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>