block: Fix SCSI block devices check
The block backend opens sysfs directory "device/block" for all available
SCSI devices to check if the target device is SCSI device or not. For
this purpose, the helper function dir_has() is called for all
directories which match /sys/class/scsi_device/*/device/block. The
dir_has() function calls opendir() for the "device/block" directories
and checks if any of files in the directory has the specified string
'entry'. If such file exists, it returns 1, and otherwise 0. However, if
the SCSI device is not a block device (e.g. SCSI enclosure),
"device/block" does not exist and opendir() and dir_has() return ENOENT.
In most cases, this ENOENT is not reported because dir_has() is called
for other SCSI block devices later and reported ENOENT is overridden
with no error for the other devices. Only when the last SCSI device to
check is not a block device, the ENOENT is reported and libzbc fails to
set up the device.

To avoid the failure, ignore ENOENT that opendir() returns for
"device/block". In this case, make dir_has() function return 0 to mean
that no file in "device/block" had the specified string 'entry'.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>