lib: Introduce zone group operations
ZBC-2/ZAC-2 add the functionality to apply zone operations not only to a
single zone at a time, but to a certain number of consecutive zones.
This is facilitated by allowing COUNT field in zone operation commands
to have values other than 0 or 1.

This commit adds support for this feature by introducing a new API
call, zbc_zone_group_op() that allows the caller to specify the number
of zones on which the zone operation in question needs to be performed.

Zone operation tools now accept an additional optional argument that
allows the user to specify the number of zones. Obviously, the number is
set to one by default.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>