lib: Add the API call to get the running libzbc version
Users can obtain the version of the installed libzbc via pkg-config,
but this only works if the library is installed to the user system by
a package manager. If libzbc is built from the source, then pkg-config
may not show the version or it may display the version of a previously
installed, obsolete version.

Add a new call to the API, zbc_version(), that returns the libzbc
package version as a string. The zbc_info tool has been updated to add
a new command line option, -V | --version, that outputs the version
string to stdout.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>