lib: Add support for XMR devices
The main focus of libzbc so far has been to provide a user interface
for HDDs using Shingled Magnetic Recording (SMR). These devices have
the entire device (except perhaps a small number of conventional zones)
use SMR ti store data.

Dynamic Hybrid Shingled Magnetic Recording (DHSMR or, more often, XMR)
is a technology that allows some portions of the device LBA space to
switch between SMR and PMR, the conventional magnetic recording method.
PMR is often called Conventional Magnetic Recording or CMR while
discussing zoned devices.

Switching zones between CMR and SMR is achieved by issuing a newly
defined set of SCSI/ATA commands to capable devices. This process is
called "zone activation" in the XMR spec, meaning that an individual
realm may be activated as either SMR or CMR.

The granularity of these conversions is device-specific, but it is
never smaller than a single zone. The most typical XMR conversion
granularity is a unit of several hundred contiguous zones called a
realm.

The zone size for XMR devices remains fixed, therefore an SMR realm
contains a larger number of zones that the corresponding CMR realm.
Since the LBA space of the device is linear and can't be stretched, the
CMR and SMR zones can't occupy the same LBA space regions. This is the
reason why XMR devices have their entire LBA space divided to large
non-overlapping regions called domains. Every domain can only have zones
of a particular type, for example, Sequential Write Required. Domain
realms can contain either active zones of that types or all zones of a
realm can be inactive. Inactive zones can't be used for reading or
writing because there is no physical storage allocated for them.

With this LBA layout, the conversion of a realm from CMR to SMR is
performed as follows: all zones of the realm in CMR domain are
inactivated and the data stored in these zones is lost. Simultaneously,
all zones of the corresponding realm in SMR domain become active and
ready for i/o. The procedure is reversed for SMR -> CMR conversions.

In the light of the above, the correct term for these conversions is
"activation". One can say that "zones of realm X are activated as SMR".

XMR is governed by Zone Domains and Zone Realms specification that
is now standardized by both T10 and T13 committees for SCSI and ATA
correspondingly and is a part of ZBC-2 and ZAC-2 command sets.

This commit adds the basic support for XMR to libzbc library code.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>
Signed-off-by: David Butterfield <David.Butterfield@wdc.com>