ata: Implement setting LBA in SET FEATURES
The existing zbc_ata_set_features() function allows the caller to set
the COUNT field in the only feature that the code currently sets,
ZBC_ATA_ENABLE_SENSE_DATA_REPORTING. Many other features defined in ATA
specification use the LBA field in SET FEATURES CDB to carry certain
input arguments for the command, and these arguments may not necessarily
be an LBA semantically.

ZAC ZD/ZR (XMR) spec uses SET FEATURES to set the number of zones to
activate and the value is placed to the LBA field when sending the
command. There is a few other uses of LBA field in SET FEATURES that are
related to XMR.

As a preparation for adding XMR support, allow setting LBA field in
SET FEATURES ATA commands.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>