treewide: Remove ZBC device emulation
When libzbc was first developed, there was virtually no hardware
available with ZBC or ZAC support. To mitigate this problem, a device
emulation (aka fake driver) has been added to libzbc. It was useful, but
it had a lot of limitations.

At the moment of this commit, a much higher quality zoned device
emulation is widely available via QEMU, nullblk Linux block driver,
TCMU Runner, etc. The fake libzbc driver has become obsolete and this
commit removes it. The utilities and test programs that are specific to
the fake driver are removed as well.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>