tests: Mark some test as "extended" and run optionally
XMR test script attempts to test various device configurations
including those that are not found in the DHSMR hardware currently
available. For example, there are no drives that use NOZSRC, or not use
FSNOZ, or support changing MAX ACTIVATION, etc. Running all these tests
makes the run time of the script excessively long and results in a lot
(> 50%) of test result records in the log marked as N/A (not
applicable).

Add a simple mechanism to the test scripts that allows to mark some of
the tests as extended. These tests are silently skipped by default. A
new test script option, -x or --run_extended_tests is introduced to
make the script run all tests including the extended ones. Mark the
most obvious candidate test cases as extended.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>