tests: Speed up execution of tests
Test scripts run noticeably slower with larger drives and the test
performance gets really sluggish when running zbc_test.sh or
zbc_dhsmr_test.sh against large DHSMR drives.

The root cause of this slowdown is the zbc_test_get_last_zone()call
that is a part of zbc_test_print_devinfo test helper program. This
function is called to fetch the info about the last zone on the device
and also to get the total number of zones. This function does that very
inefficiently - it calls zbc_list_zones for the entire device. This
results in megabytes of zone report data being transferred from the
drive. As the result, it takes zbc_print_devinfo a couple of seconds to
run against large capacity zoned drives.

Fix this inefficiency by first performing zbc_get_nr_zones() to receive
(doh) the total number of zones. The existing zbc_list_zones() is
modified to fetch only the last zone descriptor on the device. This is
done by specifying the device max_lba converted to sectors as the
starting sector of the report.

This modification results in ~50% test script run time decrease.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>