lib: Remove block backend
The block backend selection and usage has been confusing for users
of the library. Sometimes, the customers were thinking that they were
using the SCSI backend but in reality it was the block backend.

Remove the block backend functionality and make the library
passthru-only. Users can use libzbd to exercise ZBD operation via Linux
block API.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>