ata, scsi: Detect realm report layout
Some DHSMR ATA drives in the field use the "old" (or simplified) realm
report data layout. The existing code uses conditional compilation to
switch between the old and the standard layout and this is not ideal.

For ATA, check the ZAC minor version field to determine whether the
device supports the standard layout and set a device flag if it does.
Instead of conditional compilation, try the standard layout if the
flag is set and fall back to the old layout if the returned data shows
that it might be in use by the drive.

All early DHSMR devices were ATA, therefore a DHSMR SCSI device, if
found in existence, is supposed to use the standard layout. Still,
there is a caveat - an early ATA device with the old layout may be
added to the system as a SCSI drive and in this case the SAT layer
needs to be able to translate the old realm report to the standard
format, otherwise the realm report via SCSI passthrough will not
work correctly even though the ATA passthrough data from the same
device will be OK.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>