scsi: Simplify realms reporting handler
The existing code to handle REPORT REALMS is largely based on the ATA
handler. The ATA code contains some hacks to deal with early
non-standard realms data layout that is in use by some XMR devices
available in the field. The SCSI REPORT REALMS output layout was always
properly defined in ZBC-2 and there is no such thing as a non-standard
realm layout for SCSI XMR devices.

Make changes in SCSI REPORT REALMS handler to always expect the layout
defined in ZBC-2.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>