tests: Force using ATA backend for ATA XMR devices
XMR zone activation control does not work correctly with SCSI backend
when running the stock Linux kernel. This happens because the necessary
SCSI/ATA command translations are missing in the kernel.

This causes zone activation tests to be skipped when zbc_test.sh is
invoked with the default options. Adding "-ata" CLI option resolves this
problem, but this behavior can be quite confusing to the majority of
potential users.

In order to make XMR tests run smoothly out of the box, check if the drive
that is being tested is an XMR ATA and force using the ATA backend in
this case.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>