lib: ata: Fix device signature discovery in sense data
zbc_ata_classify() assumes that the sense data obtained from executing
the EXECUTE DEVICE DIAGNOSTIC command is in descriptor format. This may
not always be the case. To correctly obtain the device signature from
the sense data LBA field, the fixed data format must also be handled.

Fix this by introducing the function zbc_ata_get_signature() to properly
parse the sense data and get the device signature according to the
format of the sense data returned by the device.

Signed-off-by: Damien Le Moal <dlemoal@kernel.org>
Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>