tests: Fix NOZSRC check
The --fsnoz option should not be specified for a drive that reports
NOZSRC flag as zero. With such drives, test cases that specifically
test --fsnoz need to be skipped. However, the
zbc_test_nozsrc_check_or_NA() function that is supposed to perform
this determination incorrectly checks the presence of --fsnoz in the
test command line. As the result, it allows the test to set --fsnoz
leading to test failures.

To fix, check is_substr return code correctly.

Signed-off-by: Shinichiro Kawasaki <shinichiro.kawasaki@wdc.com>
Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>