tests: Add -skip_same option to zbc_test_zone_activate
For DH-SMR devices, the main test script, zbc_test, runs the regular
library set of tests through several activation layouts, e.g. full
SOBR, mixed SOBR/SWR and full SWR. The test scripts to configure
individual realm layouts were developed at the time when the current
protocol spec allowed activation of zones to their current type. In
this case, the activation of a zone to its current type would turn to
a NOP. Later, the spec was modified to allow these operations to
succeed. As the result, the realm layout setup code begun to fail when
running against drives with newer firmware.

To fix, add a new CLI option to zbc_test_zone_activate test program.
When specified, it will force the program to skip any zones in the
command activation range that have the same zone type as the target type
for the activation command.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>