tests: Give proper error if device in zbc_test.sh CL is omitted
If the device is not present in the command line, the script tries
to run tests using the empty string as the device name and this
leads to some unrelated cryptic errors to be displayed.

After parsing the CL, check if the device name is empty, bail out
if this is the case.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>